refactor(core): group modules into payment/ and operations/, add pre-commit hook
Move hashing, serialization, validation into payment/ module group.
Move fees into operations/ module group. Add .githooks/pre-commit
that runs biome check + build + typecheck to catch CI failures locally.
