+++ packages/core/src/index.ts
@@ -1,8 +1,6 @@
 export * from './abis/generated.js'
 export * from './config/index.js'
 export * from './errors/index.js'
-export * from './fees/index.js'
-export * from './hashing/index.js'
-export * from './serialization/index.js'
+export * from './operations/index.js'
+export * from './payment/index.js'
 export * from './types/index.js'
-export * from './validation/index.js'
