+++ packages/core/src/payment/hashing.ts
